+++ cbits/directory.c
@@ -4,6 +4,8 @@
  *
  */
 
+/* [DEPRECATED] This file may be removed in future versions. */
+
 #define INLINE
 #include "HsDirectory.h"